Login bug

+++ admin/includes/view_all_posts.php
@@ -19,6 +19,31 @@
           $delete_posts_bulk = mysqli_query($connection, $query);
           confirm($delete_posts_bulk);
         break;
+        case 'clone':
+          $query = "SELECT * FROM posts WHERE post_id = $postValueId";
+          $select_post_query = mysqli_query($connection, $query);
+          
+          while($row = mysqli_fetch_assoc($select_post_query)) {
+            $post_title         = $row['post_title'];
+            $post_category_id   = $row['post_category_id'];
+            $post_date          = $row['post_date']; 
+            $post_author        = $row['post_author'];
+            $post_status        = $row['post_status'];
+            $post_image         = $row['post_image'] ; 
+            $post_tags          = $row['post_tags']; 
+            $post_content       = $row['post_content'];
+            $post_comment_count = 0;
+
+            $query = "INSERT INTO posts(post_category_id,post_author, post_title, post_date, post_image, post_content,post_tags,post_status, post_comment_count) ";
+            $query .= "VALUES({$post_category_id}, '{$post_author}','{$post_title}',now(), '{$post_image}','{$post_content}','{$post_tags}', '{$post_status}', '{$post_comment_count}') ";
+
+            $copy_query = mysqli_query($connection, $query);
+
+            confirm($copy_query);
+          }
+
+          
+        break;
       }
     }
   }
@@ -34,6 +59,7 @@
     <option value="publish">Publish</option>
     <option value="draft">Draft</option>
     <option value="delete">Delete</option>
+    <option value="clone">Clone</option>
   </select>
 </div>
 
@@ -98,7 +124,7 @@
         <td>$post_tags</td>
         <td>$post_comment_count</td>
         <td>$post_date</td>
-        <td><a href='posts.php?delete={$post_id}'>Delete</a></td>
+        <td><a onClick=\"javascript: return confirm('Are you sure you want to delete?'); \" href='posts.php?delete={$post_id}'>Delete</a></td>
         <td><a href='posts.php?source=edit_post&p_id={$post_id}'>Update</a></td>
       </tr>";
       }
